Move Reflection.Metadata Polyfills.cs to Common/src/Polyfills/
Split into BitConverterPolyfills.cs and GuidPolyfills.cs for reuse
across libraries.
